@@ -1,4 +1,4 @@
-class AddPasswordDigestToUsers < ActiveRecord::Migration
+class AddPasswordDigestToUsers < ActiveRecord::Migration[5.0]
   def change
     add_column :users, :password_digest, :string
   end
