@@ -1,4 +1,4 @@
-class ChangeContentTypeInArticles < ActiveRecord::Migration
+class ChangeContentTypeInArticles < ActiveRecord::Migration[5.0]
   def change
     change_column :articles, :content, :text
   end
