@@ -1,4 +1,4 @@
-class AddSlugToArticles < ActiveRecord::Migration
+class AddSlugToArticles < ActiveRecord::Migration[5.0]
   def change
     add_column :articles, :slug, :string
   end
