@@ -1,4 +1,4 @@
-class AddIndexToArticlesSlug < ActiveRecord::Migration
+class AddIndexToArticlesSlug < ActiveRecord::Migration[5.0]
   def change
     add_index :articles, :slug
   end
