@@ -1,4 +1,4 @@
-class AddStatusToArticles < ActiveRecord::Migration
+class AddStatusToArticles < ActiveRecord::Migration[5.0]
   def change
     add_column :articles, :status, :integer, default: 0
   end
