@@ -1,12 +1,11 @@
 module ArticlesFilter
-  def self.filter_by_status(filter = nil)
+  extend self
+
+  def run(filter = nil)
     case filter
-    when 'all'
-      Article.all
-    when 'wip'
-      Article.wip
-    else
-      Article.published
+      when 'all' then Article.all
+      when 'wip' then Article.wip
+      else Article.published
     end
   end
 end
