@@ -1,23 +1,93 @@
+# Use this setup block to configure all options available in SimpleForm.
 SimpleForm.setup do |config|
-  config.wrappers :default, class: :input,
-                            hint_class: :field_with_hint,
-                            error_class: :field_with_errors do |b|
-    b.use :html5
-    b.use :placeholder
-    b.optional :maxlength
-    b.optional :pattern
-    b.optional :min_max
-    b.optional :readonly
-    b.use :label_input
-    b.use :hint,  wrap_with: { tag: :span, class: :hint }
-    b.use :error, wrap_with: { tag: :span, class: :error }
-  end
+  # Components used by the form builder to generate a complete input. You can remove
+  # any of them, change the order, or even add your own components to the stack.
+  # config.components = [ :placeholder, :label_input, :hint, :error ]
 
-  config.default_wrapper = :default
-  config.boolean_style = :nested
-  config.button_class = 'btn'
-  config.error_notification_tag = :div
-  config.error_notification_class = 'error_notification'
-  config.browser_validations = false
-  config.boolean_label_class = 'checkbox'
+  # Default tag used on hints.
+  # config.hint_tag = :span
+
+  # CSS class to add to all hint tags.
+  # config.hint_class = :hint
+
+  # CSS class used on errors.
+  # config.error_class = :error
+
+  # Default tag used on errors.
+  # config.error_tag = :span
+
+  # Method used to tidy up errors.
+  # config.error_method = :first
+
+  # Default tag used for error notification helper.
+  # config.error_notification_tag = :p
+
+  # CSS class to add for error notification helper.
+  # config.error_notification_class = :error_notification
+
+  # ID to add for error notification helper.
+  # config.error_notification_id = nil
+
+  # You can wrap all inputs in a pre-defined tag.
+  # config.wrapper_tag = :div
+
+  # CSS class to add to all wrapper tags.
+  # config.wrapper_class = :input
+
+  # CSS class to add to the wrapper if the field has errors.
+  # config.wrapper_error_class = :field_with_errors
+
+  # You can wrap a collection of radio/check boxes in a pre-defined tag, defaulting to none.
+  # config.collection_wrapper_tag = nil
+
+  # You can wrap each item in a collection of radio/check boxes with a tag, defaulting to span.
+  # config.item_wrapper_tag = :span
+
+  # Series of attempts to detect a default label method for collection.
+  # config.collection_label_methods = [ :to_label, :name, :title, :to_s ]
+
+  # Series of attempts to detect a default value method for collection.
+  # config.collection_value_methods = [ :id, :to_s ]
+
+  # How the label text should be generated altogether with the required text.
+  # config.label_text = lambda { |label, required| "#{required} #{label}" }
+
+  # You can define the class to use on all labels. Default is nil.
+  # config.label_class = nil
+
+  # You can define the class to use on all forms. Default is simple_form.
+  # config.form_class = :simple_form
+
+  # Whether attributes are required by default (or not). Default is true.
+  # config.required_by_default = true
+
+  # Tell browsers whether to use default HTML5 validations (novalidate option).
+  # Default is enabled.
+  # config.browser_validations = true
+
+  # Determines whether HTML5 types (:email, :url, :search, :tel) and attributes
+  # (e.g. required) are used or not. True by default.
+  # Having this on in non-HTML5 compliant sites can cause odd behavior in
+  # HTML5-aware browsers such as Chrome.
+  # config.html5 = true
+
+  # Custom mappings for input types. This should be a hash containing a regexp
+  # to match as key, and the input type that will be used when the field name
+  # matches the regexp as value.
+  # config.input_mappings = { /count/ => :integer }
+
+  # Collection of methods to detect if a file type was given.
+  # config.file_methods = [ :mounted_as, :file?, :public_filename ]
+
+  # Default priority for time_zone inputs.
+  # config.time_zone_priority = nil
+
+  # Default priority for country inputs.
+  # config.country_priority = nil
+
+  # Default size for text inputs.
+  # config.default_input_size = 50
+
+  # When false, do not use translations for labels, hints or placeholders.
+  # config.translate = true
 end
