@@ -16,7 +16,7 @@ module SessionHelper
   end
 
   def expect_login_page(page)
-    expect(page).to have_content('Login')
+    expect(page).to have_title('Login - Ralin Chimev')
     expect(page).to have_css('input[type=email]')
     expect(page).to have_css('input[type=password]')
     expect(page).to have_css('input[type=submit]')
