@@ -1,15 +1,25 @@
 require 'rails_helper'
 
 describe 'articles/index.html.haml' do
+  let(:articles) do
+    articles = create_list :article, 3
+    articles.map { |a| ArticleDecorator.new a }
+  end
+
+  before do
+    assign(:articles, articles)
+  end
+
   context 'when user is authenticated' do
+    let(:user) { create :user }
+
     before do
-      user = create :user
       create_session_for user
     end
 
     it 'displays admin links' do
-      assign(:articles, create_list(:article, 3))
       render
+
       expect(rendered).to have_link 'NEW', href: new_article_path
       expect(rendered).to have_link 'WIP', href: articles_path(status: 'wip')
       expect(rendered).to have_link 'ALL', href: articles_path(status: 'all')
@@ -17,12 +27,9 @@ describe 'articles/index.html.haml' do
   end
 
   context 'when user is guest' do
-    before do
-      assign(:articles, create_list(:article, 3))
-    end
-
     it 'does not display admin links' do
       render
+
       expect(rendered).to_not have_link 'NEW', href: new_article_path
       expect(rendered).to_not have_link 'WIP', href: articles_path(status: 'wip')
       expect(rendered).to_not have_link 'ALL', href: articles_path(status: 'all')
@@ -30,6 +37,7 @@ describe 'articles/index.html.haml' do
 
     it 'displays total number of articles' do
       render
+
       expect(rendered).to have_content '3'
     end
   end
